+++ src/Config/slackbot.config.php
@@ -6,5 +6,5 @@
  */
 
 return [
-    'version'   => '0.3.6'
+    'version'   => '0.5.0'
 ];